Tests/LibWeb/Text/expected/textarea-paste-crlf.txt

value="one\ntwo\nthree\n"
value.length=14 textLength=14 selectionStart=14
End on first line: start=4 end=4
Right from end of first line's text: start=5 end=5
